Fetch changes

+++ db/migrate/20130501105651_fix_topic_allowed_groups.rb
@@ -1,6 +1,8 @@
-class FixTopicAllowedGroups < ActiveRecord::Migration
+# frozen_string_literal: true
+
+class FixTopicAllowedGroups < ActiveRecord::Migration[4.2]
   def change
     # big oops
-    remove_column :topic_allowed_groups, :integer
+    remove_column :topic_allowed_groups, :integer if column_exists?(:topic_allowed_groups, :integer)
   end
 end
